refactor: ajusta nome da pasta com testes unitários
