+++ tests/Unit/Builders/CreditCardBuilderTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Potelo\MultiPayment\Tests\Unity\Builders;
+namespace Potelo\MultiPayment\Tests\Unit\Builders;
 
 use Illuminate\Support\Facades\Config;
 use Potelo\MultiPayment\Tests\TestCase;
